requirements.txt

--find-links https://download.pytorch.org/whl/torch_stable.html

torch==2.1.0+cu118

asposestorage==1.0.2
cdt==0.6.0
hydra-core==1.3.2
lightning==2.5.2
lingam==1.10.0
matplotlib==3.10.3
networkx==3.4.2
numpy==1.26.4
omegaconf==2.3.0
pandas==2.3.0
pyro_ppl==1.9.1
python_igraph==0.11.9
pytorch_lightning==2.5.2
PyYAML==6.0.2
rootutils==1.0.7
scikit_learn==1.7.0
scipy==1.11.4
seaborn==0.13.2
setuptools
torchvision
tqdm==4.67.1
pyarrow==20.0.0
wandb>=0.12.10
ipdb==0.13.13
cartopy==0.24.1
xarray==2025.6.1